+++ changelog.md
@@ -1,8 +1,11 @@
 ## Resourcify 1.4.0+beta.1
 
 - Add support for CurseForge
+- Settings page to configure the default pack source and some other settings
 - Convert backend to an abstract service api, allowing for easier support for more services
 - Fix background breaking when using essential (or another mod with universalcraft) in mc 1.20.5+
+- Fix crash in case of malformed html in markdown
+- Added small bisect hosting ad to the top of the browse page, this is to help support the project and can be disabled in the new settings page
 
 ----------------------------------------------------------------------------------------------------